fix distinction between stream & inbox
