+++ ui/src/index.js
@@ -18,6 +18,7 @@ import { AuthProvider } from './context/AuthContext';
 import PrivateRoutes from './utils/PrivateRoutes'
 import Login from './pages/Login/Login';
 import Profile from './pages/Profile/Profile';
+import Inbox from './pages/Inbox/Inbox';
 import Stream from './pages/Stream/Stream';
 import YourPosts from './pages/YourPosts/YourPosts';
 import PostDetail from './components/PostDetail';
@@ -35,6 +36,7 @@ root.render(
                <Route path='/' element={<App />}>
                   <Route element={<PrivateRoutes />}>
                      <Route path="/" element={<Stream/>} exact/>
+                     <Route path="/inbox" element={<Inbox />} />
                      <Route path="/posts" element={<YourPosts />} />
                      <Route path="/posts/:postid" element={<PostDetail />} />
                      <Route path="/profile" element={<Profile />} />
